Revert "Add photo vege"
