+++ app/assets/stylesheets/components/_card_vegetable.scss
@@ -37,11 +37,8 @@
 }
 .card-vege-photo{
   position: absolute;
-  width: 100px;
-  height: 80px;
   left: -14px;
   top: -20px;
-  display: cover;
 }
 
 .card-vege-icon{
