--- app/assets/stylesheets/components/_description_show.scss
@@ -1,55 +1,0 @@
-p {
-    font-size: 14px;
-  }
-
-
-.description-show{
-  z-index: 43;
-  position: fixed;
-  right: 0vw;
-  margin-top: 50px;
-  width: 700px;
-  height: auto;
-  background: rgba(250,248,245,0.9) ;
-  box-shadow: 0px 4px 4px rgba(0, 0, 0, 0.25);
-
-}
-
-.title-description{
-  margin: 20px 0px;
-  width: 22vw;
-  height: 30px;
-  background-color: rgba(0,170,37,0.25);
-  border-bottom-left-radius: 12px;
-  border-top-left-radius: 12px;
-  position: relative;
-  right: 20px;
-  h1{
-    text-align: center;
-    position: absolute;
-    bottom: -30px;
-    right: 90px;
-  }
-}
-
-.photo-description{
-  position: absolute;
-  display: cover;
-  right: 20px;
-  top: 140px;
-  width: 120px;
-  height: 90px;
-}
-
-.general-description{
-  text-align: justify;
-  font-size: 16px;
-  margin: 20px;
-  padding: 20px 120px 20px 20px;
-  background: #FFFFFF;
-  border: 1px solid rgba(0, 0, 0, 0.01);
-  box-sizing: border-box;
-  box-shadow: 0px 0px 3px rgba(0, 0, 0, 0.05);
-  border-radius: 18px;
-
-}
