+++ app/javascript/packs/application.js
@@ -1,7 +1,6 @@
 import "bootstrap";
 
-import "../plugins/description_edit";
-import "../plugins/description_show";
+import "../plugins/show_edit";
 // import { cloning } from '../plugins/garden';
 import { persisteData } from '../persiste';
