--- app/views/garden_vegetables/_show.html.erb
@@ -1,55 +1,0 @@
-<%  @garden_vegetables.each do |garden_vegetable|%>
-  <div class="description-show display-none" garden_vegetable_id=<%= garden_vegetable.id %>>
-    <i class="fas fa-minus-circle p-2"></i>
-    <div class="title-description">
-      <h1><%= garden_vegetable.vegetable.specie.capitalize %> </h1>
-    </div>
-    <div class="d-flex">
-      <p class="general-description">
-        <%= garden_vegetable.vegetable.description %>      </p>
-      <div class="vege-photo-description">
-          <%= image_tag("photo-vege/peas_photo.png", class: "photo-description") %>
-      </div>
-    </div>
-    <div class="onglet-description">
-      <nav>
-        <div class="nav nav-tabs list-inline tabs-underlined" id="nav-tab-<%=garden_vegetable.vegetable.id%>" role="tablist">
-          <% if garden_vegetable.vegetable.seeding_start?%>
-            <a class="nav-item nav-link tab-underlined active" id="nav-home-tab-<%=garden_vegetable.vegetable.id%>" data-toggle="tab" href="#nav-home-<%=garden_vegetable.vegetable.id%>" role="tab" aria-controls="nav-home" aria-selected="true"><%= image_tag("icon-seeding.png", class: "icon-action-description") %> Seeding</a>
-          <% end %>
-          <% if garden_vegetable.vegetable.planting_start?%>
-            <a class="nav-item nav-link tab-underlined" id="nav-profile-tab-<%=garden_vegetable.vegetable.id%>" data-toggle="tab" href="#nav-profile-<%=garden_vegetable.vegetable.id%>" role="tab" aria-controls="nav-profile" aria-selected="false"><%= image_tag("icon-planting.png", class: "icon-action-description") %> Planting</a>
-         <% end %>
-          <a class="nav-item nav-link tab-underlined" id="nav-contact-tab-<%=garden_vegetable.vegetable.id%>" data-toggle="tab" href="#nav-contact-<%=garden_vegetable.vegetable.id%>" role="tab" aria-controls="nav-contact" aria-selected="false"><%= image_tag("icon-harvesting.png", class: "icon-action-description") %> Harvesting</a>
-        </div>
-      </nav>
-      <div class="tab-content" id="nav-tabContent">
-        <% if garden_vegetable.vegetable.seeding_start?%>
-          <div class="tab-pane fade show active" id="nav-home-<%=garden_vegetable.vegetable.id%>" role="tabpanel" aria-labelledby="nav-home-tab">
-            <h1>Seeding</h1>
-            <h5><em>From <%= garden_vegetable.vegetable.seeding_start.strftime("%B")%> to <%=garden_vegetable.vegetable.seeding_end.strftime("%B") %></em></h3>
-            <p>
-              <%= garden_vegetable.vegetable.sedding_desription%>
-            </p>
-          </div>
-        <% end %>
-        <% if garden_vegetable.vegetable.planting_start?%>
-          <div class="tab-pane fade" id="nav-profile-<%=garden_vegetable.vegetable.id%>" role="tabpanel" aria-labelledby="nav-profile-tab">
-            <h1>Planting</h1>
-            <h5><em>From <%= garden_vegetable.vegetable.planting_start.strftime("%B")%> to <%= garden_vegetable.vegetable.planting_end.strftime("%B") %></em></h3>
-            <p>
-              <%= garden_vegetable.vegetable.planting_description%>
-            </p>
-          </div>
-        <% end %>
-        <div class="tab-pane fade" id="nav-contact-<%=garden_vegetable.vegetable.id%>" role="tabpanel" aria-labelledby="nav-contact-tab">
-           <h1>Harvesting</h1>
-          <h5><em>From <%= garden_vegetable.vegetable.harvesting_start.strftime("%B")%> to <%= garden_vegetable.vegetable.harvesting_end.strftime("%B") %></em></h3>
-          <p>
-            <%= garden_vegetable.vegetable.harvesting_description%>
-          </p>
-        </div>
-      </div>
-    </div>
-  </div>
-<% end %>
