+++ app/views/gardens/show.html.erb
@@ -2,10 +2,5 @@
   <div class="title-garden" style="position: absolute;right: 650px;top: 75px;">
     <h1><%= @garden.title %></h1>
   </div>
-  <div class="container-info">
-  <div class="vegetable-description">
-    <%= render 'garden_vegetables/show' %>
-  </div>
-</div>
   <%= render 'shared/sidebar' %>
 </div>
