+++ app/views/vegetables/_index.html.erb
@@ -5,8 +5,9 @@
 
     <div class="card-vegetable">
       <div class="vege-photo">
-        <%= image_tag("photo-vege/courgette-photo.png", class: "card-vege-photo") %>
-        <%= image_tag("courgettes.png", class: "card-vege-icon") %>
+        <%= image_tag("courgette-photo.png", class: "card-vege-photo") %>
+
+
           <div class="BlockWrapper testicule BlockWrapper--isDropzone draggable-dropzone--occupied">
             <span><div class="BlockContent"></div></span>
             <a href="#" class="Block--isDraggable vegetable" data-id="1" title="Click to plant" style="">
@@ -15,6 +16,12 @@
             </div>
             </a>v
           </div>
+
+
+
+        <%#= image_tag("courgettes.png", class: "card-vege-icon") %>
+
+
       </div>
         <h2><%= garden_vegetable.vegetable.specie.capitalize%></h2>
       <div class="d-flex justify-content-between">
