+++ app/views/vegetables/_show.html.erb
@@ -8,7 +8,7 @@
       <p class="general-description">
         <%= vegetable.description %>      </p>
       <div class="vege-photo-description">
-          <%= image_tag("photo-vege/courgette-photo.png", class: "photo-description") %>
+          <%= image_tag("courgette-photo.png", class: "photo-description") %>
       </div>
     </div>
     <div class="onglet-description">
